+++ R/helpers_ggbetween_anova_subtitles.R
@@ -15,6 +15,9 @@
 #'   `"partial_eta"` for partial eta-squared for **anova**) or `"unbiased"`
 #'   (`"g"` Hedge's *g* for **t-test**; `"partial_omega"` for partial
 #'   omega-squared for **anova**)).
+#' @param sphericity.correction Logical that decides whether to apply correction
+#'   to account for violation of sphericity in a repeated measures design ANOVA
+#'   (Default: `TRUE`).
 #' @param k Number of digits after decimal point (should be an integer)
 #'   (Default: `k = 2`).
 #' @param messages Decides whether messages references, notes, and warnings are
@@ -22,13 +25,14 @@
 #' @param ... Additional arguments.
 #' @inheritParams stats::oneway.test
 #' @inheritParams subtitle_t_parametric
-#' @inheritParams lm_effsize_standardizer
+#' @inheritParams groupedstats::lm_effsize_standardizer
 #'
 #' @importFrom dplyr select
 #' @importFrom rlang !! enquo
 #' @importFrom stats lm oneway.test na.omit
 #' @importFrom sjstats eta_sq omega_sq
 #' @importFrom ez ezANOVA
+#' @importFrom groupedstats lm_effsize_standardizer
 #'
 #' @examples
 #'
@@ -81,15 +85,36 @@ subtitle_anova_parametric <- function(data,
                                       conf.level = 0.95,
                                       nboot = 100,
                                       var.equal = FALSE,
+                                      sphericity.correction = TRUE,
                                       k = 2,
                                       messages = TRUE,
                                       ...) {
 
+  # for paired designs, variance is going to be equal across grouping levels
+  if (isTRUE(paired)) {
+    var.equal <- TRUE
+  }
+
   # number of decimal places for degree of freedom
-  if (isTRUE(var.equal) || isTRUE(paired)) {
-    k.df <- 0
+  if (isTRUE(var.equal)) {
+    if (isTRUE(paired)) {
+      if (isTRUE(sphericity.correction)) {
+        k.df2 <- k
+      } else {
+        k.df2 <- 0L
+      }
+    } else {
+      k.df2 <- 0L
+    }
   } else {
-    k.df <- k
+    k.df2 <- k
+  }
+
+  # denominator degrees of freedom
+  if (isTRUE(paired) && isTRUE(sphericity.correction)) {
+    k.df1 <- k
+  } else {
+    k.df1 <- 0L
   }
 
   # figuring out which effect size to use
@@ -165,12 +190,25 @@ subtitle_anova_parametric <- function(data,
       )
 
     # list with results
-    stats_df <-
-      list(
-        statistic = ez_df$ANOVA$F[2],
-        parameter = c(ez_df$ANOVA$DFn[2], ez_df$ANOVA$DFd[2]),
-        p.value = ez_df$ANOVA$p[2]
-      )
+    if (isTRUE(sphericity.correction)) {
+      epsilon_corr <- ez_df$`Sphericity Corrections`$GGe
+      stats_df <-
+        list(
+          statistic = ez_df$ANOVA$F[2],
+          parameter = c(
+            epsilon_corr * ez_df$ANOVA$DFn[2],
+            epsilon_corr * ez_df$ANOVA$DFd[2]
+          ),
+          p.value = ez_df$ANOVA$p[2]
+        )
+    } else {
+      stats_df <-
+        list(
+          statistic = ez_df$ANOVA$F[2],
+          parameter = c(ez_df$ANOVA$DFn[2], ez_df$ANOVA$DFd[2]),
+          p.value = ez_df$`Sphericity Corrections`$`p[GG]`[[1]]
+        )
+    }
 
     # creating a standardized dataframe with effect size and its CIs
     effsize_object <- ez_df$aov
@@ -203,13 +241,14 @@ subtitle_anova_parametric <- function(data,
   }
 
   # creating a standardized dataframe with effect size and its CIs
-  effsize_df <- lm_effsize_standardizer(
-    object = effsize_object,
-    effsize = effsize,
-    partial = partial,
-    conf.level = conf.level,
-    nboot = nboot
-  )
+  effsize_df <-
+    groupedstats::lm_effsize_standardizer(
+      object = effsize_object,
+      effsize = effsize,
+      partial = partial,
+      conf.level = conf.level,
+      nboot = nboot
+    )
 
   # preparing subtitle
   subtitle <- subtitle_template(
@@ -227,7 +266,8 @@ subtitle_anova_parametric <- function(data,
     n = sample_size,
     conf.level = conf.level,
     k = k,
-    k.parameter = k.df
+    k.parameter = k.df1,
+    k.parameter2 = k.df2
   )
 
   # message about effect size measure
@@ -507,7 +547,7 @@ subtitle_anova_robust <- function(data,
 
     # preparing the subtitle
     subtitle <-
-      base::substitute(
+      substitute(
         expr =
           paste(
             italic("F"),
@@ -526,7 +566,7 @@ subtitle_anova_robust <- function(data,
             " = ",
             n
           ),
-        env = base::list(
+        env = list(
           estimate = specify_decimal_p(x = stats_df$test[[1]], k = k),
           df1 = specify_decimal_p(x = stats_df$df1[[1]], k = k),
           df2 = specify_decimal_p(x = stats_df$df2[[1]], k = k),
@@ -571,7 +611,8 @@ subtitle_anova_robust <- function(data,
       n = sample_size,
       conf.level = conf.level,
       k = k,
-      k.parameter = k
+      k.parameter = 0L,
+      k.parameter2 = k
     )
 
     # message about effect size measure
@@ -599,22 +640,32 @@ subtitle_anova_robust <- function(data,
 #'
 #' @examples
 #' \dontrun{
+#' set.seed(123)
+#'
+#' # between-subjects ---------------------------------------
 #' # with defaults
-#' subtitle_anova_bayes(
+#' ggstatsplot::subtitle_anova_bayes(
 #'   data = ggplot2::msleep,
 #'   x = vore,
-#'   y = sleep_rem,
-#'   k = 2,
-#'   bf.prior = 0.8
+#'   y = sleep_rem
 #' )
 #'
 #' # modifying the defaults
-#' subtitle_anova_bayes(
+#' ggstatsplot::subtitle_anova_bayes(
 #'   data = ggplot2::msleep,
 #'   x = vore,
 #'   y = sleep_rem,
-#'   effsize.type = "partial_eta",
-#'   var.equal = TRUE
+#'   k = 3,
+#'   bf.prior = 0.8
+#' )
+#'
+#' # repeated measures ---------------------------------------
+#' ggstatsplot::subtitle_anova_bayes(
+#'   data = WRS2::WineTasting,
+#'   x = Wine,
+#'   y = Taste,
+#'   paired = TRUE,
+#'   k = 4
 #' )
 #' }
 #' @export
@@ -624,48 +675,9 @@ subtitle_anova_bayes <- function(data,
                                  x,
                                  y,
                                  paired = FALSE,
-                                 effsize.type = "unbiased",
-                                 partial = TRUE,
-                                 var.equal = FALSE,
                                  bf.prior = 0.707,
                                  k = 2,
                                  ...) {
-
-  # number of decimal places for degree of freedom
-  if (isTRUE(var.equal) || isTRUE(paired)) {
-    k.df <- 0
-  } else {
-    k.df <- k
-  }
-
-  # figuring out which effect size to use
-  effsize.type <- effsize_type_switch(effsize.type)
-
-  # some of the effect sizes don't work properly for paired designs
-  if (isTRUE(paired)) {
-    if (effsize.type == "unbiased") {
-      partial <- FALSE
-    } else {
-      partial <- TRUE
-    }
-  }
-
-  # preparing the subtitles with appropriate effect sizes
-  if (effsize.type == "unbiased") {
-    effsize <- "omega"
-    if (isTRUE(partial)) {
-      effsize.text <- quote(omega["p"]^2)
-    } else {
-      effsize.text <- quote(omega^2)
-    }
-  } else if (effsize.type == "biased") {
-    effsize <- "eta"
-    if (isTRUE(partial)) {
-      effsize.text <- quote(eta["p"]^2)
-    } else {
-      effsize.text <- quote(eta^2)
-    }
-  }
 
   # creating a dataframe
   data <-
@@ -689,91 +701,24 @@ subtitle_anova_bayes <- function(data,
       ) %>%
       tidyr::gather(data = ., key, value, -rowid) %>%
       dplyr::arrange(.data = ., rowid)
-
-    # sample size
-    sample_size <- length(unique(data_within$rowid))
   } else {
 
     # remove NAs listwise for between-subjects design
     data %<>%
       tidyr::drop_na(data = .)
-
-    # sample size
-    sample_size <- nrow(data)
-
-    # Welch's ANOVA run by default
-    stats_df <-
-      stats::oneway.test(
-        formula = y ~ x,
-        data = data,
-        subset = NULL,
-        na.action = na.omit,
-        var.equal = var.equal
-      )
-
-    # bayes factor results
-    bf_results <-
-      bf_oneway_anova(
-        data = data,
-        x = x,
-        y = y,
-        bf.prior = bf.prior,
-        caption = NULL,
-        output = "results"
-      )
-
-    # creating a standardized dataframe with effect size and its confidence
-    # intervals
-    effsize_df <- lm_effsize_standardizer(
-      object = stats::lm(
-        formula = y ~ x,
-        data = data,
-        na.action = na.omit
-      ),
-      effsize = effsize,
-      partial = partial
-    )
   }
 
-  # preparing the subtitle
+  # bayes factor results
   subtitle <-
-    base::substitute(
-      expr =
-        paste(
-          italic("F"),
-          "(",
-          df1,
-          ",",
-          df2,
-          ") = ",
-          estimate,
-          ", ",
-          effsize.text,
-          " = ",
-          effsize,
-          ", log"["e"],
-          "(BF"["10"],
-          ") = ",
-          bf,
-          ", ",
-          italic("r")["Cauchy"],
-          " = ",
-          bf_prior,
-          ", ",
-          italic("n"),
-          " = ",
-          n
-        ),
-      env = base::list(
-        effsize.text = effsize.text,
-        estimate = specify_decimal_p(x = stats_df$statistic[[1]], k = k),
-        df1 = stats_df$parameter[[1]],
-        df2 = specify_decimal_p(x = stats_df$parameter[[2]], k = k.df),
-        effsize = specify_decimal_p(x = effsize_df$estimate[[1]], k = k),
-        bf = specify_decimal_p(x = bf_results$log_e_bf10[[1]], k = 1),
-        bf_prior = specify_decimal_p(x = bf_results$bf.prior[[1]], k = 3),
-        n = sample_size
-      )
+    bf_oneway_anova(
+      data = data,
+      x = x,
+      y = y,
+      paired = paired,
+      bf.prior = bf.prior,
+      k = k,
+      caption = NULL,
+      output = "h1"
     )
 
   # return the subtitle